refactor: api 중복 코드 리팩터링
